refactor: split utils/handlers/helpers.py into user, callback, menu modules
Separate concerns: user validation (user.py), callback helpers (callback.py),
menu rendering with banners (menu.py). Clean up unused imports in auth.py.
