+++ handlers/core/auth.py
@@ -1,19 +1,16 @@
-import os
-import time
-from datetime import datetime
 from aiogram import Router, F
 from utils.timezone import moscow_now
 from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
-from aiogram.types import Message, CallbackQuery, InlineKeyboardMarkup, InlineKeyboardButton
+from aiogram.types import Message, CallbackQuery
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from database.db import get_user_by_tg_id, get_user_by_id, get_user_roles, get_company_by_id
+from database.db import get_user_by_tg_id, get_user_roles, get_company_by_id
 from keyboards.keyboards import get_keyboard_by_role, get_welcome_keyboard
-from states.states import AuthStates, RegistrationStates
-from utils.logger import logger, log_user_action, log_user_error
+from states.states import RegistrationStates
+from utils.logger import log_user_action, log_user_error
 from utils.bot_commands import set_bot_commands
-from utils.handlers.helpers import send_mentor_menu, send_trainee_menu
+from utils.handlers.menu import send_mentor_menu, send_trainee_menu
 
 router = Router()
 
@@ -121,162 +118,6 @@ async def cmd_login(message: Message, state: FSMContext, session: AsyncSession, 
     except Exception as e:
         log_user_error(actor.id, actor.username, "login error", e)
         await message.answer("Произошла ошибка при входе в систему. Пожалуйста, попробуй позже.")
-
-async def ensure_callback_auth(callback: CallbackQuery, state: FSMContext, session: AsyncSession) -> bool:
-    """Восстанавливает FSM state для callback-обработчиков.
-
-    callback.message.from_user — бот, поэтому check_auth/get_current_user
-    не найдут пользователя без user_id в FSM. Этот хелпер его восстанавливает.
-    Валидацию (компания, подписка) делает check_auth внутри cmd_* функций.
-    """
-    data = await state.get_data()
-    if data.get("is_authenticated") and data.get("user_id"):
-        return True
-
-    user = await get_user_by_tg_id(session, callback.from_user.id)
-    if not user or not user.is_active:
-        await callback.answer("Используй /start для входа", show_alert=True)
-        return False
-
-    roles = await get_user_roles(session, user.id)
-    primary_role = roles[0].name
-    await state.update_data(
-        user_id=user.id,
-        role=primary_role,
-        is_authenticated=True,
-        auth_time=time.time(),
-        company_id=user.company_id
-    )
-    return True
-
-
-async def check_auth(message: Message, state: FSMContext, session: AsyncSession) -> bool:
-    try:
-        data = await state.get_data()
-        is_authenticated = data.get("is_authenticated", False)
-        auth_time = data.get("auth_time", 0)
-        
-        if is_authenticated and auth_time and (time.time() - auth_time) > 86400:  # 24 часа
-            await state.clear()
-            await message.answer(
-                "👀 Ты давно не заходил\n\nДля безопасности твоя сессия завершена. Пожалуйста, обновись",
-                reply_markup=InlineKeyboardMarkup(inline_keyboard=[
-                    [InlineKeyboardButton(text="Войти заново", callback_data="login_again")]
-                ])
-            )
-            return False
-        
-        if is_authenticated:
-            user = await get_user_by_tg_id(session, message.from_user.id)
-            # Fallback: callback.message.from_user — бот, берём user_id из FSM
-            if not user and data.get("user_id"):
-                user = await get_user_by_id(session, data["user_id"])
-            if not user or not user.is_active:
-                await state.clear()
-                await message.answer("Твой аккаунт деактивирован. Обратись к администратору.")
-                return False
-            
-            # Проверка наличия компании и подписки
-            if not user.company_id:
-                await state.clear()
-                await message.answer(
-                    "❌ Ты не привязан ни к одной компании.\n\n"
-                    "Обратись к администратору."
-                )
-                return False
-            
-            # Проверка подписки компании (используем явный запрос вместо lazy loading)
-            company = await get_company_by_id(session, user.company_id)
-            if company and not company.subscribe:
-                await state.clear()
-                await message.answer(
-                    "❌ Подписка компании истекла (заморожена).\n\n"
-                    "Обратись к администратору компании для продления подписки."
-                )
-                return False
-            
-            # Проверка даты окончания подписки (по ТЗ: если finish_date прошла - доступ блокируется)
-            if company and company.finish_date and company.finish_date < moscow_now():
-                await state.clear()
-                await message.answer(
-                    "❌ Подписка компании истекла (заморожена).\n\n"
-                    "Обратись к администратору компании для продления подписки."
-                )
-                return False
-            
-            return True
-        
-        user = await get_user_by_tg_id(session, message.from_user.id)
-        # Fallback: callback.message.from_user — бот, берём user_id из FSM
-        if not user and data.get("user_id"):
-            user = await get_user_by_id(session, data["user_id"])
-
-        if not user:
-            await message.answer("Ты не зарегистрирован в системе. Используй команду /start для регистрации.")
-            return False
-
-        if not user.is_active:
-            await message.answer("Твой аккаунт деактивирован. Обратись к администратору.")
-            return False
-        
-        # Проверка наличия компании
-        if not user.company_id:
-            await message.answer(
-                "❌ Ты не привязан ни к одной компании.\n\n"
-                "Обратись к администратору."
-            )
-            return False
-        
-        # Проверка подписки компании (используем явный запрос вместо lazy loading)
-        company = await get_company_by_id(session, user.company_id)
-        if company and not company.subscribe:
-            await message.answer(
-                "❌ Подписка компании истекла (заморожена).\n\n"
-                "Обратись к администратору компании для продления подписки."
-            )
-            return False
-        
-        # Проверка даты окончания подписки (по ТЗ: если finish_date прошла - доступ блокируется)
-        if company and company.finish_date and company.finish_date < moscow_now():
-            await message.answer(
-                "❌ Подписка компании истекла (заморожена).\n\n"
-                "Обратись к администратору компании для продления подписки."
-            )
-            return False
-        
-        auto_auth_allowed = os.getenv("ALLOW_AUTO_AUTH", "true").lower() == "true"
-        if not auto_auth_allowed:
-            await message.answer("Пожалуйста, выполни команду /login для входа.")
-            return False
-        
-        roles = await get_user_roles(session, user.id)
-        
-        if not roles:
-            await message.answer("У тебя нет назначенных ролей. Обратись к рекрутеру.")
-            return False
-        
-        primary_role = roles[0].name
-        
-        await state.update_data(
-            user_id=user.id,
-            role=primary_role,
-            is_authenticated=True,
-            auth_time=message.date.timestamp(),
-            company_id=user.company_id  # КРИТИЧНО: сохраняем company_id для изоляции!
-        )
-        
-        log_user_action(
-            message.from_user.id, 
-            message.from_user.username, 
-            "auto authentication", 
-            {"role": primary_role, "user_id": user.id, "company_id": user.company_id}
-        )
-        
-        return True
-    except Exception as e:
-        log_user_error(message.from_user.id, message.from_user.username, "authentication check error", e)
-        await message.answer("Произошла ошибка при проверке авторизации. Пожалуйста, попробуй позже.")
-        return False
 
 @router.message(Command("logout"))
 async def cmd_logout(message: Message, state: FSMContext, bot):
