+++ handlers/core/common.py
@@ -5,9 +5,10 @@ from aiogram.types import Message, CallbackQuery, InlineKeyboardMarkup, InlineKe
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from database.db import get_user_by_tg_id, get_user_roles, check_user_permission
-from handlers.core.auth import check_auth
+from utils.auth.auth import check_auth
 from keyboards.keyboards import format_help_message, get_menu_by_role
-from utils.handlers.helpers import get_validated_user, cleanup_callback
+from utils.handlers.user import get_validated_user
+from utils.handlers.callback import cleanup_callback
 from utils.logger import logger, log_user_action
 from utils.messages.common import format_profile_text, get_main_menu_text, get_reload_menu_text, get_reload_inline_menu_text
 from utils.roles import get_primary_role
