+++ handlers/knowledge/knowledge_base.py
@@ -20,8 +20,9 @@ from database.db import (
     check_folder_access, get_accessible_knowledge_folders_for_user, ensure_company_id
 )
 from states.states import KnowledgeBaseStates
-from handlers.core.auth import check_auth, ensure_callback_auth
-from utils.handlers.helpers import get_current_user
+from utils.auth.auth import check_auth
+from utils.handlers.user import get_current_user
+from utils.handlers.callback import ensure_callback_auth
 from keyboards.keyboards import (
     get_knowledge_base_main_keyboard, get_knowledge_folders_keyboard,
     get_folder_created_keyboard, get_material_description_keyboard,
